requirements.txt
wordcloud==1.7.0
requests==2.22.0
requests-oauthlib==1.2.0
Pillow==6.0.0
numpy==1.19.0
Janome==0.3.10
emoji==0.5.4
